+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Next Release
+------------
+
+* Fix issue #6: `standardLogAppender` handles missing exception argument.
+
 1.0.0
 ------